requirements.txt
langchain
google-generativeai
google-api-python-client
streamlit